requirements.txt
ytmusicapi
python-dotenv
requests
